+++ NEWS.md
@@ -5,6 +5,7 @@
 * Fixed ability to change properties after deselecting the current object (#4440)
 * Fixed Properties view getting stuck with updates disabled (#4506)
 * Fixed locale-aware parsing of numbers in expression-capable spin boxes
+* Added Shift+Enter to finish polygon creation by closing the shape
 
 ### Tiled 1.12.1 (25 March 2026)